requirements.txt
fastapi
uvicorn
pydantic
pytesseract
pandas
numpy
scikit-learn
spacy
nltk
transformers
PyPDF2
requests
Pillow
python-multipart
matplotlib
seaborn
tqdm
python-dotenv
google-generativeai
torch
tensorflow
en-core-web-sm==3.8.0
